Getting the Disk Index to (mostly) work.
I have integrated the Disk Positional Index and its writer into our application. There is a hard-coded absolute path in the code that is specific to my computer, but I will try to change it to a relative path. We also need to compute document lengths to save in docWeights.bin, make another method in the Disk Positional Index to skip over positions when getting postings. and do extra tasks to get an "Excellent" for Indexing in Project Milestone 2.