Relative path implementation
I have implemented relative path implementation to make our app more portable to any computer, not just mine. The app does use a Disk Index successfully to answer term queries for the parks corpus, but not any other queries. Let us please fix that, because Prof. may grade us on those other queries for this corpus. Also, our Marvel GIF is not working for empty search results, and the app does not work with XML documents. I'm not sure if doc weights are being calculated correctly for the MLB corpus.